+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 1.2.0 - 2020-10-03
+
+### Added
+
+- zmqBridge.setRequestTimeoutValue function
+
+## Fixed
+
+- request messages queuing before the actual connection is established
+- known security vulnerabilities detected by GitHub
+
+## Updated
+
+- 'zeromq' library dependency to latest stable v5.2.0
+
 ## 1.1.0 - 2018-06-24
 
 ### Added